+++ Editor/Setting/TemplateGroupSettingEditor.cs
@@ -9,11 +9,11 @@ namespace TemplateEditor
     [CustomEditor(typeof(TemplateGroupSetting))]
     public class TemplateGroupSettingEditor : Editor
     {
-        private static readonly TemplateSettingStatus.Property[] ReplaceProperty = new TemplateSettingStatus.Property[]
+        private static readonly TemplateSettingPropertyGetter.Property[] ReplaceProperty = new TemplateSettingPropertyGetter.Property[]
         {
-            TemplateSettingStatus.Property.Path,
-            TemplateSettingStatus.Property.ScriptName,
-            TemplateSettingStatus.Property.Code,
+            TemplateSettingPropertyGetter.Property.Path,
+            TemplateSettingPropertyGetter.Property.ScriptName,
+            TemplateSettingPropertyGetter.Property.Code,
         };
 
         private TemplateGroupSetting _groupSetting;
@@ -136,25 +136,25 @@ namespace TemplateEditor
             {
                 if (GUILayout.Button("Create"))
                 {
-                    CreateScript(true);
+                    CreateScript(_replaceList, true);
                     return;
                 }
 
                 if (GUILayout.Button("No Refresh Create"))
                 {
 
-                    CreateScript(false);
+                    CreateScript(_replaceList, false);
                     return;
                 }
             }
             EditorGUILayout.EndHorizontal();
         }
 
-        private void CreateScript(bool isRefresh)
+        public void CreateScript(List<ReplaceInfo> replaceList, bool isRefresh, ProcessDictionary result = null)
         {
             foreach (var status in _statusList)
             {
-                TemplateSettingEditor.CreateScript(status, _replaceList, null, isRefresh);
+                TemplateSettingEditor.CreateScript(status, replaceList, result, isRefresh);
             }
         }