+++ Editor/Setting/TemplateSetting.cs
@@ -38,10 +38,7 @@ namespace TemplateEditor
         public string Code;
 
         [SerializeField]
-        public float CodeAreaMinHeight = 100f;
-
-        [SerializeField]
-        public float CodeAreaMaxHeight = 100f;
+        public float CodeAreaHeight = 200f;
 
         [SerializeField]
         public TemplateUtility.OverwriteType Overwrite;